Language.Nomyx.Events

Description

All the building blocks to allow rules to build events.

Documentation

onEvent :: (Typeable e, Show e) => Event e -> ((EventNumber, e) -> Nomex ()) -> Nomex EventNumber

register a callback on an event

onEvent_ :: (Typeable e, Show e) => Event e -> (e -> Nomex ()) -> Nomex EventNumber

register a callback on an event, disregard the event number

onEventOnce :: (Typeable e, Show e) => Event e -> (e -> Nomex ()) -> Nomex EventNumber

set an handler for an event that will be triggered only once

delEvent :: EventNumber -> Nomex Bool

getEvents :: NomexNE [EventInfo]

getEvent :: EventNumber -> NomexNE (Maybe EventInfo)

getIntermediateResults :: EventNumber -> NomexNE (Maybe [(PlayerNumber, SomeData)])

sendMessage :: (Typeable a, Show a) => Msg a -> a -> Nomex ()

broadcast a message that can be catched by another rule

sendMessage_ :: String -> Nomex ()

onMessage :: (Typeable m, Show m) => Msg m -> (m -> Nomex ()) -> Nomex EventNumber

subscribe on a message

onMessageOnce :: (Typeable m, Show m) => Msg m -> (m -> Nomex ()) -> Nomex EventNumber

schedule :: Schedule Freq -> (UTCTime -> Nomex ()) -> Nomex ()

on the provided schedule, the supplied function will be called

schedule_ :: Schedule Freq -> Nomex () -> Nomex ()

schedule' :: [UTCTime] -> (UTCTime -> Nomex ()) -> Nomex ()

schedule'_ :: [UTCTime] -> Nomex () -> Nomex ()

getCurrentTime :: NomexNE UTCTime

get the current time as UTCTime

oneWeek :: NominalDiffTime

duration

oneDay :: NominalDiffTime

duration

oneHour :: NominalDiffTime

duration

oneMinute :: NominalDiffTime

duration

timeEvent :: UTCTime -> Event UTCTime

Build an event firing at a specific time

messageEvent :: Typeable a => Msg a -> Event a

Build a message event, that can be intercepted by another rule this is useful for message-passing style of communication

victoryEvent :: Event VictoryInfo

Build a event firing when the victory condition is changed

playerEvent :: Player -> Event PlayerInfo

Build a event firing when a player arrives or leaves

ruleEvent :: RuleEvent -> Event RuleInfo

Build a event firing when an action is made on a rule

signalEvent :: Typeable a => Signal a -> Event a

inputFormSignal :: Typeable a => PlayerNumber -> String -> InputForm a -> Signal a

liftEvent :: NomexNE a -> Event a

Build a event firing immediatly, yelding the value of the NomexNE