Language.Nomyx.Examples

Description

This file gives a list of example rules that the players can submit. You can copy-paste them in the field Code of the web GUI. You can copy either the name of the function (i.e. "helloWorld") or its body (i.e. "outputAll_ "hello, world!""), but NOT both. Don't hesitate to get inspiration from there and create your own rules!

Documentation

nothing :: Rule

A rule that does nothing

helloWorld :: Rule

A rule that says hello to all players

accounts :: MsgVar [(PlayerNumber, Int)]

account variable name and type

createBankAccount :: Rule

Create a bank account for each players

winXEcuPerDay :: Int -> Rule

each player wins X Ecu each day you can also try with "minutly" or "monthly" instead of "daily" and everything in the "time-recurrence" package

winXEcuOnRuleAccepted :: Int -> Rule

a player wins X Ecu if a rule proposed is accepted

moneyTransfer :: Rule

a player can transfer money to another player

delRule :: RuleNumber -> Rule

delete a rule

voteWithMajority :: Rule

a majority vote, with the folowing parameters: a quorum of 2 voters is necessary for the validity of the vote the vote is assessed after every vote in case the winner is already known the vote will finish anyway after one day

king :: MsgVar PlayerNumber

makeKing :: PlayerNumber -> Rule

player pn is the king: we create a variable King to identify him, and we prefix his name with King

monarchy :: PlayerNumber -> Rule

Monarchy: only the king decides which rules to accept or reject

revolution :: PlayerNumber -> Rule

Revolution! Hail to the king! This rule suppresses the democracy (usually rules 1 and 2), installs the king and activates monarchy.

displayCurrentTime :: Rule

will display the current time (when refreshing the screen)

displayActivateTime :: Nomex ()

will display the time at which the rule as been activated

iWin :: Rule

Rule that state that you win. Good luck on having this accepted by other players ;)

returnToDemocracy :: [RuleNumber] -> Rule

Change current system (the rules passed in parameter) to absolute majority (half participants plus one)

victoryXRules :: Int -> Rule

set the victory for players having more than X accepted rules

victoryXEcu :: Int -> Rule

noGroupVictory :: Rule

Only one player can achieve victory: No group victory. Forbidding group victory usually becomes necessary when lowering the voting quorum: a coalition of players could simply force a "victory" rule and win the game.

banPlayer :: PlayerNumber -> Rule

kick a player and prevent him from returning

gameMaster :: MsgVar PlayerNumber

bravoButton :: Rule

display a button and greets you when pressed (for player 1)

enterHaiku :: Rule

displayBankAccount :: Rule

Permanently display the bank accounts

helloButton :: Rule

display a button to greet other players