Language.Nomyx.Expression

Description

This module contains the type definitions necessary to build a Nomic rule.

Documentation

type PlayerNumber = Int

type PlayerName = String

type RuleNumber = Int

type RuleName = String

type RuleDesc = String

type RuleText = String

type RuleCode = String

type EventNumber = Int

type EventName = String

type VarName = String

type Code = String

type OutputNumber = Int

type InputNumber = Int

Nomyx Expression

data Eff

Constructors

Effect

NoEffect

type Effect = Effect

type NoEffect = NoEffect

type Nomex = Exp Effect

A Nomex (Nomyx Expression) allows the players to write rules. Within the rules, you can access and modify the state of the game.

type NomexNE = Exp NoEffect

A NomexNE (Nomyx Expression No Effect) is a specialisation of the type that guarantees that the instructions will have no effects.

data Exp :: Eff -> * -> * where

Constructors

NewVar :: (Typeable a, Show a) => VarName -> a -> Nomex (Maybe (V a))

ReadVar :: (Typeable a, Show a) => V a -> NomexNE (Maybe a)

WriteVar :: (Typeable a, Show a) => V a -> a -> Nomex Bool

DelVar :: V a -> Nomex Bool

OnEvent :: (Typeable e, Show e) => Event e -> ((EventNumber, e) -> Nomex ()) -> Nomex EventNumber

DelEvent :: EventNumber -> Nomex Bool

GetEvents :: NomexNE [EventInfo]

SendMessage :: (Typeable a, Show a) => Msg a -> a -> Nomex ()

ProposeRule :: RuleInfo -> Nomex Bool

ActivateRule :: RuleNumber -> Nomex Bool

RejectRule :: RuleNumber -> Nomex Bool

AddRule :: RuleInfo -> Nomex Bool

ModifyRule :: RuleNumber -> RuleInfo -> Nomex Bool

GetRules :: NomexNE [RuleInfo]

GetPlayers :: NomexNE [PlayerInfo]

SetPlayerName :: PlayerNumber -> PlayerName -> Nomex Bool

DelPlayer :: PlayerNumber -> Nomex Bool

NewOutput :: Maybe PlayerNumber -> NomexNE String -> Nomex OutputNumber

GetOutput :: OutputNumber -> NomexNE (Maybe String)

UpdateOutput :: OutputNumber -> NomexNE String -> Nomex Bool

DelOutput :: OutputNumber -> Nomex Bool

SetVictory :: NomexNE [PlayerNumber] -> Nomex ()

CurrentTime :: NomexNE UTCTime

SelfRuleNumber :: NomexNE RuleNumber

GetRandomNumber :: Random a => (a, a) -> Nomex a

Return :: a -> Exp e a

Bind :: Exp e a -> (a -> Exp e b) -> Exp e b

ThrowError :: String -> Exp Effect a

CatchError :: Nomex a -> (String -> Nomex a) -> Nomex a

LiftEffect :: NomexNE a -> Nomex a

Simu :: Nomex a -> NomexNE Bool -> NomexNE Bool

liftEffect :: NomexNE a -> Nomex a

Variables

data V a

a container for a variable name and type

Constructors

V

Fields

varName :: String

Events

data Event a where

Composable events

Constructors

SumEvent :: Event a -> Event a -> Event a

AppEvent :: Event (a -> b) -> Event a -> Event b

PureEvent :: a -> Event a

EmptyEvent :: Event a

BindEvent :: Event a -> (a -> Event b) -> Event b

ShortcutEvents :: [Event a] -> ([Maybe a] -> Bool) -> Event [Maybe a]

SignalEvent :: Typeable a => Signal a -> Event a

LiftEvent :: NomexNE a -> Event a

data Signal a where

Signals

Constructors

Input :: PlayerNumber -> String -> InputForm a -> Signal a

Player :: Player -> Signal PlayerInfo

RuleEv :: RuleEvent -> Signal RuleInfo

Time :: UTCTime -> Signal UTCTime

Message :: Msg a -> Signal a

Victory :: Signal VictoryInfo

data SomeSignal

Type agnostic base event

Constructors

forall a . Typeable a => SomeSignal (Signal a)

data SomeData

Type agnostic result data

Constructors

forall e . (Typeable e, Show e) => SomeData e

data Player

Events parameters

Constructors

Arrive

Leave

data RuleEvent

Constructors

Proposed

Activated

Rejected

Added

Modified

Deleted

data Msg m

Constructors

Msg String

data InputForm a where

Input forms

Constructors

Text :: InputForm String

TextArea :: InputForm String

Button :: InputForm ()

Radio :: (Show a, Eq a) => [(a, String)] -> InputForm a

Checkbox :: (Show a, Eq a) => [(a, String)] -> InputForm [a]

data EventInfo

Constructors

forall e . (Typeable e, Show e) => EventInfo

Fields

_eventNumber :: EventNumber

_ruleNumber :: RuleNumber

event :: Event e

handler :: EventHandler e

_evStatus :: Status

_env :: [SignalOccurence]

data SignalAddressElem

Constructors

SumR

SumL

AppR

AppL

BindR

BindL

Shortcut

type SignalAddress = [SignalAddressElem]

data SignalData

Constructors

forall e . (Typeable e, Show e) => SignalData

Fields

signal :: Signal e

signalData :: e

data SignalOccurence

Constructors

SignalOccurence

Fields

_signalOccData :: SignalData

_signalOccAddress :: SignalAddress

type EventHandler e = (EventNumber, e) -> Nomex ()

data Status

Constructors

SActive

SDeleted

Rule

type Rule = Nomex ()

Type of a rule function.

data RuleInfo

An informationnal structure about a rule

Constructors

RuleInfo

Fields

_rNumber :: RuleNumber

_rName :: RuleName

_rDescription :: String

_rProposedBy :: PlayerNumber

_rRuleCode :: Code

_rRule :: Rule

_rStatus :: RuleStatus

_rAssessedBy :: Maybe RuleNumber

data RuleStatus

the status of a rule.

Constructors

Active

Pending

Reject

Player

data PlayerInfo

informations on players

Constructors

PlayerInfo

Fields

_playerNumber :: PlayerNumber

_playerName :: String

_playAs :: Maybe PlayerNumber

Victory

data VictoryInfo

Constructors

VictoryInfo

Fields

_vRuleNumber :: RuleNumber

_vCond :: NomexNE [PlayerNumber]

Miscellaneous

partial :: String -> Nomex (Maybe a) -> Nomex a

concatMapM :: Monad m => (a -> m [b]) -> [a] -> m [b]

signalOccAddress :: Lens SignalOccurence SignalAddress

signalOccData :: Lens SignalOccurence SignalData

env :: Lens EventInfo [SignalOccurence]

evStatus :: Lens EventInfo Status

ruleNumber :: Lens EventInfo RuleNumber

eventNumber :: Lens EventInfo EventNumber

playAs :: Lens PlayerInfo (Maybe PlayerNumber)

playerName :: Lens PlayerInfo String

playerNumber :: Lens PlayerInfo PlayerNumber

rAssessedBy :: Lens RuleInfo (Maybe RuleNumber)

rStatus :: Lens RuleInfo RuleStatus

rRule :: Lens RuleInfo Rule

rRuleCode :: Lens RuleInfo Code

rProposedBy :: Lens RuleInfo PlayerNumber

rDescription :: Lens RuleInfo String

rName :: Lens RuleInfo RuleName

rNumber :: Lens RuleInfo RuleNumber