Language.Nomyx.Inputs

Description

All the building blocks to allow rules to get inputs. for example, you can create a button that will display a message like this: do void $ onInputButton_ "Click here:" (const $ outputAll_ "Bravo!") 1

Documentation

data InputForm a where

Input forms

Constructors

Text :: InputForm String

TextArea :: InputForm String

Button :: InputForm ()

Radio :: (Show a, Eq a) => [(a, String)] -> InputForm a

Checkbox :: (Show a, Eq a) => [(a, String)] -> InputForm [a]

inputRadio :: (Eq c, Show c, Typeable c) => PlayerNumber -> String -> [(c, String)] -> Event c

event based on a radio input choice

inputText :: PlayerNumber -> String -> Event String

event based on a text input

inputCheckbox :: (Eq c, Show c, Typeable c) => PlayerNumber -> String -> [(c, String)] -> Event [c]

event based on a checkbox input

inputButton :: PlayerNumber -> String -> Event ()

event based on a button

inputTextarea :: PlayerNumber -> String -> Event String

event based on a text area

onInputRadio :: (Typeable a, Eq a, Show a) => String -> [a] -> (EventNumber -> a -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

triggers a choice input to the user. The result will be sent to the callback

onInputRadio_ :: (Typeable a, Eq a, Show a) => String -> [a] -> (a -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

the same, disregard the event number

onInputRadioOnce :: (Typeable a, Eq a, Show a) => String -> [a] -> (a -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

the same, suppress the event after first trigger

inputRadio' :: (Eq c, Show c, Typeable c) => PlayerNumber -> String -> [c] -> Event c

onInputText :: String -> (EventNumber -> String -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

triggers a string input to the user. The result will be sent to the callback

onInputText_ :: String -> (String -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

asks the player pn to answer a question, and feed the callback with this data.

onInputTextOnce :: String -> (String -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

asks the player pn to answer a question, and feed the callback with this data.

onInputCheckbox :: (Typeable a, Eq a, Show a) => String -> [(a, String)] -> (EventNumber -> [a] -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputCheckbox_ :: (Typeable a, Eq a, Show a) => String -> [(a, String)] -> ([a] -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputCheckboxOnce :: (Typeable a, Eq a, Show a) => String -> [(a, String)] -> ([a] -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputButton :: String -> (EventNumber -> () -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputButton_ :: String -> (() -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputButtonOnce :: String -> (() -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputTextarea :: String -> (EventNumber -> String -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputTextarea_ :: String -> (String -> Nomex ()) -> PlayerNumber -> Nomex EventNumber

onInputTextareaOnce :: String -> (String -> Nomex ()) -> PlayerNumber -> Nomex EventNumber