Language.Nomyx.Outputs

Description

All the building blocks to allow rules to produce outputs. for example, you can display a message like this: do outputAll_ "hello, world!"

Documentation

type OutputNumber = Int

newOutput :: Maybe PlayerNumber -> NomexNE String -> Nomex OutputNumber

outputs a message to one player, dynamic version

newOutput_ :: Maybe PlayerNumber -> String -> Nomex OutputNumber

outputs a message to one player, static version

outputAll :: NomexNE String -> Nomex OutputNumber

output a message to all players

outputAll_ :: String -> Nomex ()

output a constant message to all players

getOutput :: OutputNumber -> NomexNE (Maybe String)

get an output by number

getOutput_ :: OutputNumber -> Nomex String

get an output by number, partial version

updateOutput :: OutputNumber -> NomexNE String -> Nomex Bool

update an output

delOutput :: OutputNumber -> Nomex Bool

delete an output

displayVar :: (Typeable a, Show a) => Maybe PlayerNumber -> MsgVar a -> (Maybe a -> NomexNE String) -> Nomex OutputNumber

displayVar' :: (Typeable a, Show a) => Maybe PlayerNumber -> MsgVar a -> (a -> NomexNE String) -> Nomex OutputNumber

displaySimpleVar :: (Typeable a, Show a) => Maybe PlayerNumber -> MsgVar a -> String -> Nomex OutputNumber

displayArrayVar :: (Typeable a, Show a, Typeable i, Show i) => Maybe PlayerNumber -> ArrayVar i a -> String -> Nomex OutputNumber