Language.Nomyx.Players

Description

All the building blocks to allow rules to manage players. for example, you can change the name of player 1 with: do void $ modifyPlayerName 1 ("King " ++)

Documentation

type PlayerNumber = Int

type PlayerName = String

data PlayerInfo

informations on players

Constructors

PlayerInfo

Fields

_playerNumber :: PlayerNumber

_playerName :: String

_playAs :: Maybe PlayerNumber

data Player

Events parameters

Constructors

Arrive

Leave

playerEvent :: Player -> Event PlayerInfo

Build a event firing when a player arrives or leaves

playerNumber :: Lens PlayerInfo PlayerNumber

playerName :: Lens PlayerInfo String

getPlayers :: NomexNE [PlayerInfo]

get all the players

getPlayer :: PlayerNumber -> NomexNE (Maybe PlayerInfo)

Get a specific player

getPlayerName :: PlayerNumber -> NomexNE (Maybe PlayerName)

Get the name of a player

getPlayerName' :: PlayerNumber -> NomexNE PlayerName

Get the name of a player, his number if not found

setPlayerName :: PlayerNumber -> PlayerName -> Nomex Bool

Set the name of a player

modifyPlayerName :: PlayerNumber -> (PlayerName -> PlayerName) -> Nomex Bool

getPlayersNumber :: NomexNE Int

Get the total number of players

getAllPlayerNumbers :: NomexNE [PlayerNumber]

Get all the players number

delPlayer :: PlayerNumber -> Nomex Bool

Remove the player from the game (kick)

forEachPlayer :: (PlayerNumber -> Nomex ()) -> (PlayerNumber -> Nomex ()) -> (PlayerNumber -> Nomex ()) -> Nomex (EventNumber, EventNumber)

perform an action for each current players, new players and leaving players returns the event numbers for arriving players and leaving players

forEachPlayer_ :: (PlayerNumber -> Nomex ()) -> Nomex (EventNumber, EventNumber)

perform the same action for each players, including new players returns the event numbers for arriving players and leaving players

createValueForEachPlayer :: forall a. (Typeable a, Show a, Eq a) => a -> MsgVar [(PlayerNumber, a)] -> Nomex (EventNumber, EventNumber)

create a value initialized for each players manages players joining and leaving

createValueForEachPlayer_ :: MsgVar [(PlayerNumber, Int)] -> Nomex (EventNumber, EventNumber)

create a value initialized for each players initialized to zero manages players joining and leaving

getValueOfPlayer :: forall a. (Typeable a, Show a, Eq a) => PlayerNumber -> MsgVar [(PlayerNumber, a)] -> NomexNE (Maybe a)

modifyValueOfPlayer :: (Eq a, Show a, Typeable a) => PlayerNumber -> MsgVar [(PlayerNumber, a)] -> (a -> a) -> Nomex Bool

modifyAllValues :: (Eq a, Show a, Typeable a) => MsgVar [(PlayerNumber, a)] -> (a -> a) -> Nomex ()

showPlayer :: PlayerNumber -> NomexNE String

show a player name based on his number

getProposerNumber :: NomexNE PlayerNumber

get the player number of the proposer of the rule

getProposerNumber_ :: Nomex PlayerNumber

setVictory :: NomexNE [PlayerNumber] -> Nomex ()

set victory to a list of players

giveVictory :: PlayerNumber -> Nomex ()

give victory to one player

getRandomNumber :: Random a => (a, a) -> Nomex a

get a random number uniformly distributed in the closed interval [lo,hi] resets the number generator