Language.Nomyx.Rules

Description

Basic rules building blocks. for example, you can suppress rule 1 with: do suppressRule 1

Documentation

type RuleNumber = Int

type RuleCode = String

data RuleEvent

Constructors

Proposed

Activated

Rejected

Added

Modified

Deleted

data RuleStatus

the status of a rule.

Constructors

Active

Pending

Reject

type MetaRule = RuleInfo -> NomexNE Bool

A meta rule is a rule that can juge the legality of another rule.

activateRule :: RuleNumber -> Nomex Bool

activate a rule: change its state to Active and execute it

activateRule_ :: RuleNumber -> Nomex ()

rejectRule :: RuleNumber -> Nomex Bool

reject a rule: change its state to Suppressed and suppresses all its environment (events, variables, inputs, victory) the rule can be activated again later

rejectRule_ :: RuleNumber -> Nomex ()

getRules :: NomexNE [RuleInfo]

getActiveRules :: NomexNE [RuleInfo]

getRule :: RuleNumber -> NomexNE (Maybe RuleInfo)

getRulesByNumbers :: [RuleNumber] -> NomexNE [RuleInfo]

getRuleFuncs :: NomexNE [Nomex ()]

addRule :: RuleInfo -> Nomex Bool

add a rule to the game, it will have to be activated

addRule_ :: RuleInfo -> Nomex ()

addRule' :: RuleName -> Rule -> RuleCode -> String -> Nomex RuleNumber

add a rule to the game as described by the parameters

getFreeRuleNumber :: NomexNE RuleNumber

suppressRule :: RuleNumber -> Nomex Bool

suppressRule_ :: RuleNumber -> Nomex ()

suppressAllRules :: Nomex Bool

modifyRule :: RuleNumber -> RuleInfo -> Nomex Bool

autoActivate :: Nomex ()

This rule will activate automatically any new rule.

activateOrRejectRule :: RuleInfo -> Bool -> Nomex ()

activate or reject a rule

simulate :: Nomex a -> NomexNE Bool -> NomexNE Bool

simulate the execution of rule "sim" and then run rule "test" over the result

metaruleVar :: MsgVar [(String, MetaRule)]

The meta rules are stored in a list variable

createMetaruleVar :: Nomex ()

create the meta rule variable

addMetarule :: MetaRule -> String -> Nomex ()

add a new metarule to the list

testWithMetaRules :: RuleInfo -> NomexNE Bool

use the list of meta rules to juge a new rule

displayMetarules :: Nomex ()

legal :: MetaRule

A rule will be always legal

illegal :: MetaRule

A rule will be always illegal

noPlayPlayer :: PlayerNumber -> MetaRule

Player p cannot propose any more rules

immutableRule :: RuleNumber -> MetaRule

rule number rn cannot be deleted by any incoming rule we simulate the execution of an incoming rule to make sure it doesn't delete the immutable rule

autoDelete :: Nomex ()

a rule can autodelete itself (generaly after having performed some actions)

eraseAllRules :: PlayerNumber -> Nomex Bool

All rules from player p are erased:

getSelfRuleNumber :: NomexNE RuleNumber

allows a rule to retrieve its own number (for auto-deleting for example)

getSelfRule :: NomexNE RuleInfo

onRuleProposed :: (RuleInfo -> Nomex ()) -> Nomex ()

sets a callback called for each rule proposed

showRule :: Show a => a -> Exp Effect ()