Language.Nomyx.Variables

Description

All the building blocks to allow rules to build variables. for example, you can create a variable with: do newMsgVar_ MyMoney (0::Int)

Documentation

data V a

a container for a variable name and type

Constructors

V

Fields

varName :: String

data VEvent a

a MsgVar is a variable with a message attached, allowing to trigger registered functions anytime the var if modified

Constructors

VUpdated a

VDeleted

data MsgVar a

Constructors

MsgVar

Fields

message :: Msg (VEvent a)

variable :: V a

newVar :: (Typeable a, Show a) => VarName -> a -> Nomex (Maybe (V a))

variable creation

newVar_ :: (Typeable a, Show a) => VarName -> a -> Nomex (V a)

newVar' :: (Typeable a, Show a) => V a -> a -> Nomex Bool

readVar :: (Typeable a, Show a) => V a -> NomexNE (Maybe a)

variable reading

readVar_ :: (Typeable a, Show a) => V a -> Nomex a

writeVar :: (Typeable a, Show a) => V a -> a -> Nomex Bool

variable writing

modifyVar :: (Typeable a, Show a) => V a -> (a -> a) -> Nomex Bool

modify a variable using the provided function

delVar :: V a -> Nomex Bool

delete variable

newMsgVar :: (Typeable a, Show a) => VarName -> a -> Nomex (Maybe (MsgVar a))

newMsgVar_ :: (Typeable a, Show a) => VarName -> a -> Nomex (MsgVar a)

newMsgVar' :: (Typeable a, Show a) => MsgVar a -> a -> Nomex Bool

readMsgVar :: (Typeable a, Show a) => MsgVar a -> NomexNE (Maybe a)

readMsgVar_ :: (Typeable a, Show a) => MsgVar a -> Nomex a

writeMsgVar :: (Typeable a, Show a) => MsgVar a -> a -> Nomex Bool

modifyMsgVar :: (Typeable a, Show a) => MsgVar a -> (a -> a) -> Nomex Bool

delMsgVar :: (Typeable a, Show a) => MsgVar a -> Nomex Bool

msgVar :: String -> MsgVar a

newMsgVarOnEvent :: (Typeable a, Show a, Eq a) => VarName -> a -> (VEvent a -> Nomex ()) -> Nomex (Maybe (MsgVar a))

create a new MsgVar and register callback in case of change (update, delete)

onMsgVarEvent :: (Typeable a, Show a) => MsgVar a -> (VEvent a -> Nomex ()) -> Nomex EventNumber

onMsgVarChange

Arguments

:: (Typeable a, Show a, Eq a)

=> MsgVar a
the MsgVar

-> (a -> Nomex b)
callback on creation (called immediatly)

-> (a -> b -> Nomex ())
callback on update

-> (b -> Nomex ())
callback on delete

-> Nomex EventNumber
event number generated for update and delete

adds a callback for each of the MsgVar events: Create, Update, Delete

onMsgVarDelete :: (Typeable a, Show a) => MsgVar a -> Nomex () -> Nomex EventNumber

getMsgVarMessage :: (Typeable a, Show a) => MsgVar a -> NomexNE (Msg (VEvent a))

get the messsage triggered when the array is filled

getMsgVarData :: (Typeable a, Show a) => MsgVar a -> NomexNE (Maybe a)

get the association array

getMsgVarData_ :: (Typeable a, Show a) => MsgVar a -> Nomex a

getMsgVarName :: (Typeable a, Show a) => MsgVar a -> String

type ArrayVar i a = MsgVar [(i, Maybe a)]

ArrayVar is an indexed array with a signal attached triggered at every change. | each indexed elements starts empty (value=Nothing).

newArrayVar :: (Typeable a, Show a, Typeable i, Show i) => VarName -> [i] -> Nomex (Maybe (ArrayVar i a))

initialize an empty ArrayVar

newArrayVar_ :: (Typeable a, Show a, Typeable i, Show i) => VarName -> [i] -> Nomex (ArrayVar i a)

newArrayVar' :: (Typeable a, Show a, Eq a, Typeable i, Show i, Eq i) => VarName -> [i] -> (VEvent [(i, Maybe a)] -> Nomex ()) -> Nomex (Maybe (ArrayVar i a))

initialize an empty ArrayVar, registering a callback that will be triggered at every change

newArrayVarOnce :: (Typeable a, Show a, Eq a, Typeable i, Show i, Ord i) => VarName -> [i] -> (VEvent [(i, Maybe a)] -> Nomex ()) -> Nomex (Maybe (ArrayVar i a))

initialize an empty ArrayVar, registering a callback. the ArrayVar will be deleted when full

cleanOnFull :: (Typeable a, Show a, Eq a, Typeable i, Show i, Ord i) => ArrayVar i a -> Nomex ()

isFullArrayVar_ :: (Typeable a, Show a, Typeable i, Show i, Ord i) => ArrayVar i a -> NomexNE Bool

putArrayVar :: (Typeable a, Show a, Eq a, Typeable i, Show i, Eq i, Ord i) => ArrayVar i a -> i -> a -> Nomex Bool

store one value and the given index. If this is the last filled element, the registered callbacks are triggered.

putArrayVar_ :: (Typeable a, Show a, Eq a, Typeable i, Show i, Ord i) => ArrayVar i a -> i -> a -> Nomex ()