Language.Nomyx.Vote

Description

Voting system

Documentation

type AssessFunction = VoteStats -> Maybe Bool

a vote assessing function (such as unanimity, majority...)

data VoteStats

the vote statistics, including the number of votes per choice, the number of persons called to vote, and if the vote is finished (timeout or everybody voted)

Constructors

VoteStats

Fields

voteCounts :: Map Bool Int

nbParticipants :: Int

voteFinished :: Bool

unanimityVote :: Nomex ()

vote at unanimity every incoming rule

callVoteRule :: AssessFunction -> NominalDiffTime -> RuleInfo -> Nomex ()

call a vote on a rule for every players, with an assessing function and a delay

callVoteRule' :: AssessFunction -> UTCTime -> RuleInfo -> Nomex ()

finishVote :: AssessFunction -> RuleInfo -> [(PlayerNumber, Maybe Bool)] -> Nomex ()

actions to do when the vote is finished

callVote :: AssessFunction -> UTCTime -> String -> RuleNumber -> ([(PlayerNumber, Maybe Bool)] -> Nomex ()) -> Nomex ()

call a vote for every players, with an assessing function, a delay and a function to run on the result

voteWith :: UTCTime -> AssessFunction -> String -> Event [(PlayerNumber, Maybe Bool)]

vote with a function able to assess the ongoing votes. | the vote can be concluded as soon as the result is known.

singleVote :: String -> PlayerNumber -> Event Bool

unanimity :: AssessFunction

assess the vote results according to a unanimity

majority :: AssessFunction

assess the vote results according to an absolute majority (half voters plus one)

majorityWith :: Int -> AssessFunction

assess the vote results according to a majority of x (in %)

numberVotes :: Int -> AssessFunction

assess the vote results according to a fixed number of positive votes

withQuorum :: AssessFunction -> Int -> AssessFunction

adds a quorum to an assessing function

getVoteStats :: [Maybe Bool] -> Bool -> VoteStats

counts :: (Eq a, Ord a) => [a] -> [(a, Int)]

voteQuota :: Int -> VoteStats -> Maybe Bool

Compute a result based on a quota of positive votes. the result can be positive if the quota if reached, negative if the quota cannot be reached anymore at that point, or still pending.

nbVoters :: VoteStats -> Int

number of people that voted if the voting is finished, total number of people that should vote otherwise

voted :: VoteStats -> Int

notVoted :: VoteStats -> Int

displayVote :: EventNumber -> RuleNumber -> Nomex ()

getVotes :: [PlayerNumber] -> [(PlayerNumber, Bool)] -> [(PlayerNumber, Maybe Bool)]

getBooleanResult :: (PlayerNumber, SomeData) -> (PlayerNumber, Bool)

showOnGoingVote :: [(PlayerNumber, Maybe Bool)] -> RuleNumber -> NomexNE String

showFinishedVote :: RuleNumber -> Bool -> [(PlayerNumber, Maybe Bool)] -> NomexNE String

showVote :: (PlayerNumber, Maybe Bool) -> NomexNE (String, String)

showChoice :: Maybe Bool -> String